이미지 기능 추가 (#17)
* Fix : 예외 컨트롤러 같은 클래스 사용 수정

* Feat : 게시물,매물,후기,신고 기능 추가, 페이지네이션 적용

* Feat : 게시물,후기 페이징 API 작성, 내가 올린 매물, 찜한 매물 조회 API 주가

* Feat : report관련 admin api 추가

* Feat : 이미지 등록,조회 api 추가

+++ dankan/src/main/java/com/dankan/domain/embedded/RoomStructure.java
@@ -26,7 +26,7 @@ public class RoomStructure {
     @Column(name = "structure",nullable = false,length = 24,columnDefinition = "varchar")
     private String structure;
 
-    @Column(name = "room_size",nullable = false,columnDefinition = "int")
+    @Column(name = "room_size",nullable = false,columnDefinition = "double")
     private Double roomSize;
 
     @Column(name = "real_room_size",nullable = false,columnDefinition = "double")
